Relocated CSS files into associated component directories
