+++ client/src/components/App/App.jsx
@@ -10,7 +10,7 @@ import Register from "../Register/Register.jsx";
 import Login from "../Login/Login.jsx";
 import { AuthProvider } from "react-auth-kit";
 import { UserProvider, useUser } from "../UserProvider.jsx";
-import "../../styles/App.css";
+import "./App.css";
 import Calendar from "../Calendar/Calendar.jsx";
 
 const App = () => {
