+++ client/src/components/CohortRender/CohortDetails.css
