+++ client/src/components/LeftColumn/LeftColumn.css
